feat(update): add terminal animations for update and install flows
Add polished terminal animations using indicatif and console crates:

- Braille dot spinners during update checks and installation phases
- Styled version transition boxes with emoji
- Staged progress feedback (download, extract, replace, verify)
- Enhanced install.sh with ASCII banner and animated stages

Implements the design from docs/plans/2026-01-29-update-animations-design.md

+++ src/docker/containers.rs
@@ -190,8 +190,8 @@ impl From<bollard::models::ContainerSummary> for ContainerSummary {
             .filter_map(|p| {
                 Some(crate::core::PortMapping {
                     ip: p.ip.map(|s| s.to_string()),
-                    private_port: p.private_port as u16,
-                    public_port: p.public_port.map(|p| p as u16),
+                    private_port: p.private_port,
+                    public_port: p.public_port.map(|p| p),
                     protocol: p
                         .typ
                         .map(|t| format!("{:?}", t))
